# Optimize io.cache ##io