--- a/libr/include/r_vector.h
+++ b/libr/include/r_vector.h
@@ -76,6 +76,12 @@ static inline bool r_vector_empty(const RVector *vec) {
 
 R_API void r_vector_clear(RVector *vec);
 
+// returns the length of the vector
+static inline size_t r_vector_len(const RVector *vec) {
+	r_return_val_if_fail (vec, 0);
+	return vec->len;
+}
+
 // returns a pointer to the offset inside the array where the element of the index lies.
 static inline void *r_vector_index_ptr(RVector *vec, size_t index) {
 	r_return_val_if_fail (vec && index < vec->capacity, NULL);