--- a/libr/io/io_cache.c
+++ b/libr/io/io_cache.c
@@ -3,6 +3,7 @@
 // TODO: implement a more intelligent way to store cached memory
 
 #include "r_io.h"
+#include <r_skyline.h>
 
 #if 0
 #define CACHE_CONTAINER(x) container_of ((RBNode*)x, RCache, rb)
@@ -22,36 +23,32 @@ static void cache_item_free(RIOCache *cache) {
 }
 
 R_API bool r_io_cache_at(RIO *io, ut64 addr) {
-	RIOCache *c;
-	RListIter *iter;
-	r_list_foreach (io->cache, iter, c) {
-		if (r_itv_contain (c->itv, addr)) {
-			return true;
-		}
-	}
-	return false;
+	r_return_val_if_fail (io, false);
+	return r_skyline_contains (&io->cache_skyline, addr);
 }
 
 R_API void r_io_cache_init(RIO *io) {
-	io->cache = r_list_newf ((RListFree)cache_item_free);
+	r_pvector_init (&io->cache, (RPVectorFree)cache_item_free);
+	r_skyline_init (&io->cache_skyline);
 	io->buffer = r_cache_new ();
 	io->cached = 0;
 }
 
-R_API void r_io_cache_fini (RIO *io) {
-	r_list_free (io->cache);
+R_API void r_io_cache_fini(RIO *io) {
+	r_pvector_fini (&io->cache);
+	r_skyline_fini (&io->cache_skyline);
 	r_cache_free (io->buffer);
-	io->cache = NULL;
 	io->buffer = NULL;
 	io->cached = 0;
 }
 
 R_API void r_io_cache_commit(RIO *io, ut64 from, ut64 to) {
-	RListIter *iter;
+	void **iter;
 	RIOCache *c;
 	RInterval range = (RInterval){from, to - from};
-	r_list_foreach (io->cache, iter, c) {
+	r_pvector_foreach (&io->cache, iter) {
 		// if (from <= c->to - 1 && c->from <= to - 1) {
+		c = *iter;
 		if (r_itv_overlap (c->itv, range)) {
 			int cached = io->cached;
 			io->cached = 0;
@@ -68,37 +65,47 @@ R_API void r_io_cache_commit(RIO *io, ut64 from, ut64 to) {
 
 R_API void r_io_cache_reset(RIO *io, int set) {
 	io->cached = set;
-	r_list_purge (io->cache);
+	r_pvector_clear (&io->cache);
+	r_skyline_clear (&io->cache_skyline);
 }
 
 R_API int r_io_cache_invalidate(RIO *io, ut64 from, ut64 to) {
 	int invalidated = 0;
-	RListIter *iter, *tmp;
+	void **iter;
 	RIOCache *c;
 	RInterval range = (RInterval){from, to - from};
-	r_list_foreach_prev_safe (io->cache, iter, tmp, c) {
+	r_pvector_foreach_prev (&io->cache, iter) {
+		c = *iter;
 		if (r_itv_overlap (c->itv, range)) {
 			int cached = io->cached;
 			io->cached = 0;
 			r_io_write_at (io, r_itv_begin (c->itv), c->odata, r_itv_size (c->itv));
 			io->cached = cached;
 			c->written = false;
-			r_list_delete (io->cache, iter);
+			r_pvector_remove_data (&io->cache, c);
 			invalidated++;
 		}
+	}
+	r_skyline_clear (&io->cache_skyline);
+	r_pvector_foreach_prev (&io->cache, iter) {
+		c = *iter;
+		r_skyline_add (&io->cache_skyline, c->itv, c);
 	}
 	return invalidated;
 }
 
 R_API int r_io_cache_list(RIO *io, int rad) {
-	int i, j = 0;
-	RListIter *iter;
+	size_t i, j = 0;
+	void **iter;
 	RIOCache *c;
+	PJ *pj;
 	if (rad == 2) {
-		io->cb_printf ("[");
+		pj = pj_new ();
+		pj_a (pj);
 	}
-	r_list_foreach (io->cache, iter, c) {
-		const int dataSize = r_itv_size (c->itv);
+	r_pvector_foreach (&io->cache, iter) {
+		c = *iter;
+		const ut64 dataSize = r_itv_size (c->itv);
 		if (rad == 1) {
 			io->cb_printf ("wx ");
 			for (i = 0; i < dataSize; i++) {
@@ -111,20 +118,20 @@ R_API int r_io_cache_list(RIO *io, int rad) {
 			}
 			io->cb_printf ("\n");
 		} else if (rad == 2) {
-			io->cb_printf ("{\"idx\":%d,\"addr\":%"PFMT64d",\"size\":%d,",
-				j, r_itv_begin (c->itv), dataSize);
-			io->cb_printf ("\"before\":\"");
-		  	for (i = 0; i < dataSize; i++) {
-				io->cb_printf ("%02x", c->odata[i]);
-			}
-			io->cb_printf ("\",\"after\":\"");
-		  	for (i = 0; i < dataSize; i++) {
-				io->cb_printf ("%02x", c->data[i]);
-			}
-			io->cb_printf ("\",\"written\":%s}%s", c->written
-				? "true": "false", iter->n? ",": "");
+			pj_o (pj);
+			pj_kn (pj, "idx", j);
+			pj_kn (pj, "addr", r_itv_begin (c->itv));
+			pj_kn (pj, "size", dataSize);
+			char *hex = r_hex_bin2strdup (c->odata, dataSize);
+			pj_ks (pj, "before", hex);
+			free (hex);
+			hex = r_hex_bin2strdup (c->data, dataSize);
+			pj_ks (pj, "after", hex);
+			free (hex);
+			pj_kb (pj, "written", c->written);
+			pj_end (pj);
 		} else if (rad == 0) {
-			io->cb_printf ("idx=%d addr=0x%08"PFMT64x" size=%d ", j, r_itv_begin (c->itv), dataSize);
+			io->cb_printf ("idx=%"PFMTSZu" addr=0x%08"PFMT64x" size=%"PFMT64u" ", j, r_itv_begin (c->itv), dataSize);
 			for (i = 0; i < dataSize; i++) {
 				io->cb_printf ("%02x", c->odata[i]);
 			}
@@ -137,7 +144,10 @@ R_API int r_io_cache_list(RIO *io, int rad) {
 		j++;
 	}
 	if (rad == 2) {
-		io->cb_printf ("]\n");
+		pj_end (pj);
+		char *json = pj_drain (pj);
+		io->cb_printf ("%s", json);
+		free (json);
 	}
 	return false;
 }
@@ -168,29 +178,41 @@ R_API bool r_io_cache_write(RIO *io, ut64 addr, const ut8 *buf, int len) {
 		io->cachemode = cm;
 	}
 	memcpy (ch->data, buf, len);
-	r_list_append (io->cache, ch);
+	r_pvector_push (&io->cache, ch);
+	r_skyline_add (&io->cache_skyline, ch->itv, ch);
 	REventIOWrite iow = { addr, buf, len };
 	r_event_send (io->event, R_EVENT_IO_WRITE, &iow);
 	return true;
 }
 
 R_API bool r_io_cache_read(RIO *io, ut64 addr, ut8 *buf, int len) {
+	r_return_val_if_fail (io, false);
+	RSkyline *skyline = &io->cache_skyline;
+	const RSkylineItem *iter = r_skyline_get_item_intersect (skyline, addr, len);
+	if (!iter) {
+		return false;
+	}
+	const RSkylineItem *last = (RSkylineItem *)skyline->v.a + skyline->v.len;
 	bool covered = false;
-	RListIter *iter;
-	RIOCache *c;
-	RInterval range = (RInterval){ addr, len };
-	r_list_foreach (io->cache, iter, c) {
-		if (r_itv_overlap (c->itv, range)) {
-			const ut64 begin = r_itv_begin (c->itv);
-			if (addr < begin) {
-				int l = R_MIN (addr + len - begin, r_itv_size (c->itv));
-				memcpy (buf + begin - addr, c->data, l);
-			} else {
-				int l = R_MIN (r_itv_end (c->itv) - addr, len);
-				memcpy (buf, c->data + addr - begin, l);
-			}
-			covered = true;
+	while (iter != last) {
+		const ut64 begin = r_itv_begin (iter->itv);
+		const st64 addr_offset = begin - addr;
+		const ut64 buf_offset = addr_offset > 0 ? addr_offset : 0;
+		const ut64 cur_addr = addr + buf_offset;
+		const ut64 left = len - buf_offset;
+		if (begin > cur_addr + left) {
+			break;
 		}
+		RIOCache *cache = iter->user;
+		const ut64 cache_shift = addr_offset < 0 ? -addr_offset : 0;
+		const ut64 cache_offset = begin - r_itv_begin (cache->itv) + cache_shift;
+		const ut64 read = R_MIN (left, r_itv_size (iter->itv) - cache_shift);
+		memcpy (buf + buf_offset, cache->data + cache_offset, read);
+		covered = true;
+		if (left - read <= 0) {
+			break;
+		}
+		iter++;
 	}
 	return covered;
 }