--- a/libr/util/meson.build
+++ b/libr/util/meson.build
@@ -76,6 +76,7 @@ r_util_sources = [
   'utf16.c',
   'utf32.c',
   'vector.c',
+  'skyline.c',
   'w32-sys.c',
   'zip.c',
   'protobuf.c',