--- a/test/unit/test_io.c
+++ b/test/unit/test_io.c
@@ -1,6 +1,55 @@
 #include <r_io.h>
 #include "minunit.h"
 
+bool test_r_io_cache(void) {
+	RIO *io = r_io_new ();
+	r_io_open (io, "malloc://15", R_PERM_RW, 0);
+	r_io_write (io, "ZZZZZZZZZZZZZZZ", 15);
+	r_io_cache_init (io);
+	mu_assert_false (r_io_cache_at (io, 0), "Cache shouldn't exist at 0");
+	mu_assert_false (r_io_cache_at (io, 10), "Cache shouldn't exist at 10");
+	mu_assert_true (r_io_cache_write (io, 0, "AAAAA", 5), "Cache write at 0 failed");
+	mu_assert_true (r_io_cache_write (io, 10, "BBBBB", 5), "Cache write at 10 failed");
+	mu_assert_true (r_io_cache_at (io, 0), "Cache should exist at 0 (beggining of cache)");
+	mu_assert_true (r_io_cache_at (io, 4), "Cache should exist at 4 (end of cache)");
+	mu_assert_false (r_io_cache_at (io, 8), "Cache shouldn't exist at 8 (between 2 caches)");
+	mu_assert_true (r_io_cache_at (io, 12), "Cache should exist at 12 (middle of cache)");
+	ut8 buf[15];
+	memset (buf, 'Z', sizeof (buf));
+	mu_assert_true (r_io_cache_read (io, 0, buf, sizeof (buf)), "Cache read failed");
+	mu_assert_memeq (buf, "AAAAAZZZZZBBBBB", sizeof (buf), "Cache read doesn't match expected output");
+	memset (buf, 'Z', sizeof (buf));
+	mu_assert_true (r_io_cache_write (io, 0, "CC", 2), "Overlapped cache write at 0 failed");
+	mu_assert_true (r_io_cache_write (io, 4, "DD", 2), "Overlapped cache write at 4 failed");
+	mu_assert_true (r_io_cache_write (io, 8, "EEE", 3), "Cache write at 4 failed");
+	mu_assert_true (r_io_cache_read (io, 0, buf, 2), "Cache read at 0 failed");
+	mu_assert_true (r_io_cache_read (io, 2, buf + 2, 2), "Cache read at 2 failed");
+	mu_assert_true (r_io_cache_read (io, 4, buf + 4, 2), "Cache read at 4 failed");
+	mu_assert_true (r_io_cache_read (io, 6, buf + 6, 2), "Cache read at 6 failed");
+	mu_assert_true (r_io_cache_read (io, 8, buf + 8, 3), "Cache read at 8 failed");
+	mu_assert_true (r_io_cache_read (io, 11, buf + 11, 4), "Cache read at 11 failed");
+	mu_assert_memeq (buf, "CCAADDZZEEEBBBB", sizeof (buf), "Cache read doesn't match expected output");
+	mu_assert_true (r_io_cache_write (io, 0, "FFFFFFFFFFFFFFF", 15), "Cache write failed");
+	mu_assert_true (r_io_cache_read (io, 0, buf, sizeof (buf)), "Cache read failed");
+	mu_assert_memeq (buf, "FFFFFFFFFFFFFFF", sizeof (buf), "Cache read doesn't match expected output");
+	r_io_read_at (io, 0, buf, sizeof (buf));
+	mu_assert_memeq (buf, "ZZZZZZZZZZZZZZZ", sizeof (buf), "IO read without cache doesn't match expected output");
+	io->cached = R_PERM_R;
+	r_io_read_at (io, 0, buf, sizeof (buf));
+	mu_assert_memeq (buf, "FFFFFFFFFFFFFFF", sizeof (buf), "IO read with cache doesn't match expected output");
+	r_io_cache_invalidate (io, 6, 1);
+	memset (buf, 'Z', sizeof (buf));
+	r_io_read_at (io, 0, buf, sizeof (buf));
+	mu_assert_memeq (buf, "CCAADDZZEEEBBBB", sizeof (buf), "IO read after cache invalidate doesn't match expected output");
+	r_io_cache_commit (io, 0, 15);
+	memset (buf, 'Z', sizeof (buf));
+	io->cached = 0;
+	r_io_read_at (io, 0, buf, sizeof (buf));
+	mu_assert_memeq (buf, "CCAADDZZEEEBBBB", sizeof (buf), "IO read after cache commit doesn't match expected output");
+	r_io_free (io);
+	mu_end;
+}
+
 bool test_r_io_mapsplit (void) {
 	RIO *io = r_io_new ();
 	io->va = true;
@@ -213,6 +262,7 @@ bool test_r_io_priority2(void) {
 }
 
 int all_tests() {
+	mu_run_test(test_r_io_cache);
 	mu_run_test(test_r_io_mapsplit);
 	mu_run_test(test_r_io_mapsplit2);
 	mu_run_test(test_r_io_mapsplit3);