<!-- Filling this template is mandatory -->

**Your checklist for this pull request**
- [x] I've read the [guidelines for contributing](https://github.com/radareorg/radare2/blob/master/DEVELOPERS.md) to this repository
- [x] I made sure to follow the project's [coding style](https://github.com/radareorg/radare2/blob/master/DEVELOPERS.md#code-style)
- [ ] I've added tests that prove my fix is effective or that my feature works (if possible)
- [ ] I've updated the documentation and the [radare2 book](https://github.com/radareorg/radare2book) with the relevant information (if needed)

**Detailed description**

<!-- Explain the **details** for making this change. Is a new feature implemented? What existing problem does the pull request solve? How does the pull request solve these issues? Please provide enough information so that others can review your pull request. -->

* Use RPVector instead of RList
* Add cache skyline for faster reads
* Use PJ in `wcj`

Relocs killed r2's performance

#### Before

`r2 -e bin.dbginfo=0 -e io.cache=1 -qq libr/asm/libr_asm.so  93.41s user 0.19s system 100% cpu 1:33.59 total`

#### After

`r2 -ebin.dbginfo=0 -eio.cache=1 -qq libr/asm/libr_asm.so  0.97s user 0.16s system 100% cpu 1.120 total`

#### Without `-e io.cache=1`

`r2 -e bin.dbginfo=0 -qq libr/asm/libr_asm.so  0.77s user 0.14s system 99% cpu 0.914 total`

**Test plan**

<!-- What steps should the reviewer take to test your pull request? Demonstrate the code is solid. Example: The exact commands you ran and their output, screenshots/videos. This is your time to re-check that everything works and that you covered all the edge cases -->

Tests pass